requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
absl-py=1.3.0=pypi_0
astor=0.8.1=pypi_0
awscli=1.27.29=pypi_0
blas=1.0=mkl
botocore=1.29.29=pypi_0
brotli=1.0.9=h5eee18b_7
brotli-bin=1.0.9=h5eee18b_7
bzip2=1.0.8=h7b6447c_0
ca-certificates=2022.10.11=h06a4308_0
cachetools=4.2.4=pypi_0
cairo=1.14.12=h8948797_3
certifi=2022.9.24=py37h06a4308_0
charset-normalizer=2.1.1=pypi_0
colorama=0.4.4=pypi_0
cudatoolkit=10.1.243=h6bb024c_0
cudnn=7.6.5=cuda10.1_0
cycler=0.11.0=pyhd3eb1b0_0
dbus=1.13.18=hb2f20db_0
docutils=0.16=pypi_0
expat=2.4.9=h6a678d5_0
ffmpeg=4.0=hcdf2ecd_0
flit-core=3.6.0=pyhd3eb1b0_0
fontconfig=2.14.1=hef1e5e3_0
fonttools=4.25.0=pyhd3eb1b0_0
freeglut=3.0.0=hf484d3e_5
freetype=2.12.1=h4a9f257_0
gast=0.2.2=pypi_0
giflib=5.2.1=h7b6447c_0
glib=2.63.1=h5a9c865_0
google-auth=1.35.0=pypi_0
google-auth-oauthlib=0.4.6=pypi_0
google-pasta=0.2.0=pypi_0
graphite2=1.3.14=h295c915_1
grpcio=1.51.1=pypi_0
gst-plugins-base=1.14.0=hbbd80ab_1
gstreamer=1.14.0=hb453b48_1
h5py=3.7.0=pypi_0
harfbuzz=1.8.8=hffaf4a1_0
hdf5=1.10.2=hba1933b_1
icu=58.2=he6710b0_3
idna=3.4=pypi_0
importlib-metadata=5.1.0=pypi_0
imutils=0.5.4=pypi_0
intel-openmp=2021.4.0=h06a4308_3561
jasper=2.0.14=hd8c5072_2
jmespath=1.0.1=pypi_0
jpeg=9e=h7f8727e_0
keras-applications=1.0.8=pypi_0
keras-preprocessing=1.1.0=pypi_0
kiwisolver=1.4.2=py37h295c915_0
lcms2=2.12=h3be6417_0
lerc=3.0=h295c915_0
libbrotlicommon=1.0.9=h5eee18b_7
libbrotlidec=1.0.9=h5eee18b_7
libbrotlienc=1.0.9=h5eee18b_7
libdeflate=1.8=h7f8727e_5
libedit=3.1.20210910=h7f8727e_0
libffi=3.2.1=hf484d3e_1007
libgcc-ng=11.2.0=h1234567_1
libgfortran-ng=7.5.0=ha8ba4b0_17
libgfortran4=7.5.0=ha8ba4b0_17
libglu=9.0.0=hf484d3e_1
libgomp=11.2.0=h1234567_1
libopencv=3.4.2=hb342d67_1
libopus=1.3.1=h7b6447c_0
libpng=1.6.37=hbc83047_0
libstdcxx-ng=11.2.0=h1234567_1
libtiff=4.4.0=hecacb30_2
libvpx=1.7.0=h439df22_0
libwebp=1.2.4=h11a3e52_0
libwebp-base=1.2.4=h5eee18b_0
libxcb=1.15=h7f8727e_0
libxml2=2.9.14=h74e7548_0
lxml=4.9.1=pypi_0
lz4-c=1.9.3=h295c915_1
markdown=3.4.1=pypi_0
markupsafe=2.1.1=pypi_0
matplotlib=3.5.3=py37h06a4308_0
matplotlib-base=3.5.3=py37hf590b9c_0
mkl=2021.4.0=h06a4308_640
mkl-service=2.4.0=py37h7f8727e_0
mkl_fft=1.3.1=py37hd3c417c_0
mkl_random=1.2.2=py37h51133e4_0
munkres=1.1.4=py_0
ncurses=6.3=h5eee18b_3
numpy=1.18.5=pypi_0
oauthlib=3.2.2=pypi_0
opencv=3.4.2=py37h6fd60c2_1
opencv-contrib-python=4.6.0.66=pypi_0
opencv-python=4.6.0.66=pypi_0
openssl=1.0.2u=h7b6447c_0
opt-einsum=3.3.0=pypi_0
packaging=21.3=pyhd3eb1b0_0
pandas=1.1.5=pypi_0
pcre=8.45=h295c915_0
pillow=9.3.0=pypi_0
pip=22.3.1=py37h06a4308_0
pixman=0.40.0=h7f8727e_1
protobuf=3.20.0=pypi_0
py-opencv=3.4.2=py37hb342d67_1
pyasn1=0.4.8=pypi_0
pyasn1-modules=0.2.8=pypi_0
pyparsing=3.0.9=py37h06a4308_0
pyqt=5.9.2=py37h05f1152_2
python=3.7.0=h6e4f718_3
python-dateutil=2.8.2=pyhd3eb1b0_0
pytz=2022.6=pypi_0
pyyaml=5.4.1=pypi_0
qt=5.9.6=h8703b6f_2
readline=7.0=h7b6447c_5
requests=2.28.1=pypi_0
requests-oauthlib=1.3.1=pypi_0
rsa=4.7.2=pypi_0
s3transfer=0.6.0=pypi_0
scipy=1.7.3=pypi_0
seaborn=0.12.1=pypi_0
setuptools=65.5.0=py37h06a4308_0
sip=4.19.8=py37hf484d3e_0
six=1.16.0=pyhd3eb1b0_1
sqlite=3.33.0=h62c20be_0
tensorboard=2.1.1=pypi_0
tensorflow-estimator=2.1.0=pypi_0
tensorflow-gpu=2.1.2=pypi_0
termcolor=2.1.1=pypi_0
tk=8.6.12=h1ccaba5_0
tornado=6.2=py37h5eee18b_0
tqdm=4.64.1=pypi_0
typing_extensions=4.4.0=py37h06a4308_0
urllib3=1.26.13=pypi_0
utils=1.0.1=pypi_0
werkzeug=2.2.2=pypi_0
wheel=0.37.1=pyhd3eb1b0_0
wrapt=1.14.1=pypi_0
xz=5.2.8=h5eee18b_0
zipp=3.11.0=pypi_0
zlib=1.2.13=h5eee18b_0
zstd=1.5.2=ha4553b6_0
